lots of work

+++ app/src/main/java/com/ruskaof/myt/domain/repository/PlansRepository.kt
@@ -6,5 +6,5 @@ import kotlinx.coroutines.flow.Flow
 interface PlansRepository {
     fun getPlans(): Flow<List<PlanDbo>>
     suspend fun writePlan(plan: PlanDbo)
-    suspend fun removePlan(id: Int)
+    suspend fun removePlan(id: Long)
 }
