+++ app/src/main/java/com/ruskaof/myt/domain/use_case/plans/RemovePlanUseCase.kt
@@ -1,13 +1,10 @@
 package com.ruskaof.myt.domain.use_case.plans
 
-import com.ruskaof.myt.common.Resource
 import com.ruskaof.myt.domain.repository.PlansRepository
-import kotlinx.coroutines.flow.flow
-import java.lang.Exception
 import javax.inject.Inject
 
 class RemovePlanUseCase @Inject constructor(
     private val plansRepository: PlansRepository
 ) {
-    suspend operator fun invoke(id: Int) = plansRepository.removePlan(id)
+    suspend operator fun invoke(id: Long) = plansRepository.removePlan(id)
 }
